refactor(devtools-view): Move data visualization components into their own directory (#15558)
Also moves `Tree` into `utility-components`.

This PR is generally aimed at just reducing the number of modules in the
root of `components`.
