+++ packages/tools/devtools/devtools-view/src/components/data-visualization/FluidTreeView.tsx
@@ -3,11 +3,13 @@
  * Licensed under the MIT License.
  */
 import React from "react";
+
 import { HasContainerId, FluidObjectTreeNode } from "@fluid-experimental/devtools-core";
-import { Tree } from "./Tree";
+
+import { Tree } from "../utility-components";
+import { HasLabel } from "./CommonInterfaces";
 import { TreeDataView } from "./TreeDataView";
 import { TreeHeader } from "./TreeHeader";
-import { HasLabel } from "./CommonInterfaces";
 
 /**
  * {@link TreeView} input props.
