+++ packages/tools/devtools/devtools-view/src/components/data-visualization/FluidValueView.tsx
@@ -5,9 +5,11 @@
 import React from "react";
 // eslint-disable-next-line import/no-internal-modules
 import { TreeItem } from "@fluentui/react-components/unstable";
+
 import { FluidObjectValueNode } from "@fluid-experimental/devtools-core";
-import { TreeHeader } from "./TreeHeader";
+
 import { HasLabel } from "./CommonInterfaces";
+import { TreeHeader } from "./TreeHeader";
 
 /**
  * {@link ValueView} input props.
