+++ packages/tools/devtools/devtools-view/src/components/data-visualization/TreeView.tsx
@@ -4,11 +4,13 @@
  */
 
 import React from "react";
+
 import { HasContainerId, VisualTreeNode } from "@fluid-experimental/devtools-core";
-import { Tree } from "./Tree";
+
+import { Tree } from "../utility-components";
+import { HasLabel } from "./CommonInterfaces";
 import { TreeDataView } from "./TreeDataView";
 import { TreeHeader } from "./TreeHeader";
-import { HasLabel } from "./CommonInterfaces";
 
 /**
  * {@link TreeView} input props.
