+++ packages/tools/devtools/devtools-view/src/components/index.ts
@@ -7,21 +7,15 @@
 /* eslint-disable no-restricted-syntax */
 
 export * from "./client-data-views";
+export * from "./data-visualization";
 
 export * from "./AudienceView";
 export * from "./ContainerDevtoolsView";
 export * from "./ContainerHistoryView";
 export * from "./ContainerSummaryView";
-export * from "./FluidHandleView";
-export * from "./FluidTreeView";
-export * from "./FluidValueView";
 export * from "./LandingView";
 export * from "./Menu";
 export * from "./TelemetryView";
-export * from "./TreeView";
-export * from "./UnknownDataView";
-export * from "./UnknownFluidObjectView";
-export * from "./ValueView";
 export * from "./Waiting";
 
 /* eslint-enable no-restricted-syntax */
