+++ packages/tools/devtools/devtools-view/src/components/utility-components/index.ts
@@ -3,9 +3,5 @@
  * Licensed under the MIT License.
  */
 
-// TODO: re-enable rule and fix violations before merging into main
-/* eslint-disable no-restricted-syntax */
-
-export * from "./Accordion";
-
-/* eslint-enable no-restricted-syntax */
+export { Accordion, AccordionProps } from "./Accordion";
+export { Tree, TreeProps } from "./Tree";
